% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{bnp.ar}
\alias{bnp.ar}
\title{Fitting a Bayesian Nonparametric AutoRegressive model of order q - AR(q)}
\usage{
bnp.ar(
  Y,
  q,
  initializer,
  X = matrix(0, ncol = 0, nrow = 0),
  iter = 10000,
  burn_in = 5000,
  thin = 10,
  conf_level = 0.05,
  sigma = 0,
  plot = F
)
}
\arguments{
\item{Y}{Time series}

\item{q}{Order of the AutoRegressive model AR(q)}

\item{initializer}{list of initial values - can be obtained as the output of the function initializer()}

\item{X}{Covariates of interest (by default, the function fits only the autoregressive part)}

\item{iter}{number of iterations to consider}

\item{burn_in}{number of initial sampled valued to discard}

\item{thin}{thinning value}

\item{conf_level}{confidence level of the credible intervals}

\item{sigma}{parameter to regulate the Pitman-Yor process - default: sigma=0 (Dirichlet process case).}

\item{cluster}{boolean value indicating whether the cluster analysis is desired}
}
\description{
Fitting a Bayesian Nonparametric AutoRegressive model of order q - AR(q)
}
\keyword{ar}
\keyword{ar,}
\keyword{autoregressive}
\keyword{model}
\keyword{model,}
