% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{bnp.GP}
\alias{bnp.GP}
\title{Fitting a Gaussian Process with the Residual Structure explicited in Ascolani, Ghidini (2021+)}
\usage{
bnp.GP(
  X,
  Y,
  initializer,
  iter,
  Xstar,
  burn_in = 0,
  thin = 1,
  Verbose = T,
  log_rate = T,
  sigma = 0
)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response variable}

\item{initializer}{list of initial values - can be obtained as the output of the function initializer()}

\item{iter}{number of iterations to consider}

\item{Xstar}{grid to predict}

\item{burn_in}{number of initial sampled valued to discard}

\item{thin}{thinning value}

\item{sigma}{parameter to regulate the Pitman-Yor process - default: sigma=0 (Dirichlet process case).}
}
\description{
Fitting a Gaussian Process with the Residual Structure explicited in Ascolani, Ghidini (2021+)
}
\keyword{gaussian}
\keyword{gp}
\keyword{model,}
\keyword{process,}
