% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{bnp.lm}
\alias{bnp.lm}
\title{Fitting a Bayesian Nonparametric linear model}
\usage{
bnp.lm(
  X,
  Y,
  initializer,
  cluster = F,
  iter = 10000,
  burn_in = 5000,
  thin = 10,
  conf_level = 0.05,
  sigma = 0
)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response variable}

\item{initializer}{list of initial values - can be obtained as the output of the function initializer()}

\item{cluster}{boolean value indicating whether the cluster analysis is desired}

\item{iter}{number of iterations to consider}

\item{burn_in}{number of initial sampled valued to discard}

\item{thin}{thinning value}

\item{conf_level}{confidence level of the credible intervals}

\item{sigma}{parameter to regulate the Pitman-Yor process - default: sigma=0 (Dirichlet process case).}
}
\description{
This function allows you to ofit the linear model defined in Ascolani, Ghidini (year)
}
\keyword{linear}
\keyword{lm}
\keyword{model,}
