% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{calcSigma}
\alias{calcSigma}
\title{Fitting a Gaussian Process with the Residual Structure explicited in Ascolani, Ghidini (2021+)}
\usage{
calcSigma(X1, X2, l = 1)
}
\arguments{
\item{X1}{first matrix for the computation of the kernel}

\item{X2}{second matrix for the computation of the kernel}

\item{l}{parameter of the kernel}
}
\description{
Fitting a Gaussian Process with the Residual Structure explicited in Ascolani, Ghidini (2021+)
}
\keyword{covariance}
\keyword{kernel,}
