% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{coverage}
\alias{coverage}
\title{Coverage}
\usage{
coverage(real_value, intervals)
}
\arguments{
\item{real_value}{true value of the parameters of interest}

\item{intervals}{credible or confidence intervals}
}
\description{
This function allows you to compute the coverage of the credible (or confidence) intervals
}
\keyword{coverage,}
\keyword{diagnostic}
