% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{credible_intervals}
\alias{credible_intervals}
\title{Credible Intervals}
\usage{
credible_intervals(parameters, level = 0.05)
}
\arguments{
\item{parameters}{empirical posterior distribution of the parameters of interest}

\item{level}{credible level (default: level=0.05)}
}
\description{
This function allows you to compute credible intervals of the linear regression coefficients
}
\keyword{credible}
\keyword{diagnostic}
\keyword{intervals,}
