% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{designmatrix.ar}
\alias{designmatrix.ar}
\title{This function defines a design matrix, to estimate the AR(q) model}
\usage{
designmatrix.ar(X, Y, q)
}
\arguments{
\item{X}{Covariates of interest}

\item{Y}{Time Series of interest}

\item{q}{Order of the AutoRegressive model AR(q)}
}
\description{
This function defines a design matrix, to estimate the AR(q) model
}
