% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{model_selection}
\alias{model_selection}
\title{Model Selection}
\usage{
model_selection(X, Y, sigma2, cutoff_level = 0.75)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response variable}

\item{cutoff_level}{threshold to include a variable (default: cutoff_level = 0.75)}
}
\description{
This function allows you to perform model selection
}
\keyword{diagnostic}
\keyword{model}
\keyword{selection,}
