% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{sampler_beta}
\alias{sampler_beta}
\title{Sample the coefficients of the linear model}
\usage{
sampler_beta(mu, tau1, tau2, comp, Y, X, B0, b0)
}
\arguments{
\item{mu}{absolute value of the mean of the components of the residual mixture}

\item{tau1}{variance of the first component of the residual mixture}

\item{tau2}{variance of the second component of the residual mixture}

\item{comp}{vector of dimension n, with possible entries +1 or -1, denoting the element of the mixture from which the residual has been sampled from}

\item{Y}{response variable}

\item{X}{design matrix}

\item{B0}{covariance matrix for the gaussian prior of beta}

\item{b0}{mean vector for the gaussian prior of beta}
}
\description{
This function allows you to sample from the posterior distribution of the parameters beta, as explained in Ascolani, Ghidini (2021)
}
\keyword{beta_sampler}
\keyword{sampler_beta,}
