% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{sampler_gamma_prob}
\alias{sampler_gamma_prob}
\title{Compute the distribution of gamma}
\usage{
sampler_gamma_prob(X, Y, beta, mu, tau1, tau2, log_rate = T)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response variable}

\item{beta}{coefficients of the linear model}

\item{mu}{absolute value of the mean of the components of the residual mixture}

\item{tau1}{variance of the first component of the residual mixture}

\item{tau2}{variance of the second component of the residual mixture}

\item{log_rate}{compute all the probabilities in log-scale (log_rate=T)}
}
\description{
This function allows you to compute the probability distribution of gamma, as explained in Ascolani, Ghidini (year)
}
\keyword{sampler_gamma_prob}
