% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{sampler_triplet_blockedGS}
\alias{sampler_triplet_blockedGS}
\title{Blocked Gibbs sampler algorithm}
\usage{
sampler_triplet_blockedGS(Y, epsilon, comp, p, Z, mu0, sigma0)
}
\arguments{
\item{epsilon}{residuals}

\item{comp}{vector of dimension n, with possible entries +1 or -1, denoting the element of the mixture from which the residual has been sampled from}

\item{p}{distribution of the N labels}

\item{Z}{matrix containing the unique values of the triplets (Z1, Z2, Z3), each one associated with a label in 1:N}

\item{mu0}{hyperparameter for the baseline measure (mean of the Truncated Normal)}

\item{sigma0}{hyperparameter for the baseline measure (variance of the Truncated Normal)}

\item{X}{design matrix}
}
\description{
This function allows you to sample an instance of (Z1, Z2, Z3) from the posterior distribution,
exploiting the Blocked Gibbs Sampler algorithm as explained in Ascolani, Ghidini (year)
}
\keyword{blocked}
\keyword{gibbs}
\keyword{sampler}
\keyword{sampler_triplet_blockedGS,}
