% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpResiduals_functions.R
\name{p_sampler}
\alias{p_sampler}
\title{Sampling the distribution of the labels}
\usage{
p_sampler(K, theta, N)
}
\arguments{
\item{K}{set of classifications labels sampled in a previous iteration}

\item{theta}{concentration parameter}

\item{N}{number of elements in the truncation of the PY process}
}
\description{
This function allows you to sample labels k=1, ..., N from the distribution defined in Ishwaran and James, 2001
}
\keyword{p_sampler}
